Update rusqlite and some other stuff (#13)
rusqlite 0.10 didn't seem to build any more... I arbitrarily perturbed things
until it seemed to build again?

+++ src/database.rs
@@ -77,7 +77,7 @@ impl<'a> DbRow<'a> {
     pub fn  get<T: FromSql>(&mut self, idx: u8) -> T {
         match *self {
             DbRow::Postgres(ref mut row) => row.get(idx as usize),
-            DbRow::Sqlite(ref mut row) => row.get(idx as i32),
+            DbRow::Sqlite(ref mut row) => row.get(idx as usize),
         }
     }
 }
